+++ backend/src/test/java/com/bia/dev_bank/serviceTest/CardServiceTest.java
@@ -64,7 +64,7 @@ public class CardServiceTest {
                 new ArrayList<>(),
                 new ArrayList<>(),
                 new ArrayList<>(),
-                500.0,
+                BigDecimal.valueOf(500.0),
                 LocalDate.now()
         );
